Refactor controller files
