+++ packages/server/src/api/controllers/row/handlers/ExternalRequest.ts
@@ -22,15 +22,17 @@ import {
   getPrimaryDisplay,
   isRowId,
   isSQL,
-} from "../../../integrations/utils"
-import { getDatasourceAndQuery } from "./utils"
-import { FieldTypes } from "../../../constants"
+} from "../../../../integrations/utils"
+import { getDatasourceAndQuery } from "../utils"
+import { FieldTypes } from "../../../../constants"
 import { processObjectSync } from "@budibase/string-templates"
 import { cloneDeep } from "lodash/fp"
-import { processDates, processFormulas } from "../../../utilities/rowProcessor"
+import {
+  processDates,
+  processFormulas,
+} from "../../../../utilities/rowProcessor"
 import { db as dbCore } from "@budibase/backend-core"
-import sdk from "../../../sdk"
-import { isEditableColumn } from "../../../sdk/app/tables/validation"
+import sdk from "../../../../sdk"
 
 export interface ManyRelationship {
   tableId?: string
